abstracting permissions functions
